% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ErrorMetrics.R
\name{se}
\alias{se}
\alias{mse}
\alias{rmse}
\alias{sse}
\alias{rse}
\alias{rrse}
\alias{ae}
\alias{mae}
\alias{mdae}
\alias{ape}
\alias{mape}
\title{Error Metrics}
\usage{
se(x, y)

mse(x, y, ...)

rmse(x, y, ...)

sse(x, y)

rse(x, y)

rrse(x, y)

ae(x, y)

mae(x, y, ...)

mdae(x, y, ...)

ape(x, y)

mape(x, y, ...)
}
\arguments{
\item{x}{Actual/observed values}

\item{y}{Predicted values}

\item{...}{Other arguments passed to inner mean and median functions.}
}
\description{
The functions calculate error metrics between two numeric vectors.

\code{mse}: Mean squared difference.

\code{rmse}: Root Mean Squared Difference.

\code{sse}: Sum of the squared difference.

\code{rse}: Relative squared error.

\code{rrse} Root relative squared error.

\code{ae}: Absolute difference.

\code{ae}: Mean Absolute difference.

\code{ae}: Median Absolute difference.

\code{ape}: Elementwise absoulte percent difference.

\code{ape}: Average elementwise absoulte percent difference.
}
\details{
\code{se}: (Squared Error) elementwise squared difference.
}
